+++ packages/spacecat-shared-data-access/test/it/sentiment-topic/sentiment-topic.test.js
@@ -15,6 +15,7 @@ import chaiAsPromised from 'chai-as-promised';
 
 import { getDataAccess } from '../util/db.js';
 import { seedDatabase } from '../util/seed.js';
+import { IT_HOOK_TIMEOUT } from '../util/util.js';
 
 use(chaiAsPromised);
 
@@ -35,7 +36,7 @@ describe('SentimentTopic IT', function () {
 
   // eslint-disable-next-line prefer-arrow-callback
   before(async function () {
-    this.timeout(10000);
+    this.timeout(IT_HOOK_TIMEOUT);
     sampleData = await seedDatabase();
 
     const dataAccess = getDataAccess();